removed updatedUserDto and used UserDto instead, updated userservice, userrespository and usercontroller as needed, added casting for userRole on meeting repository

+++ KlockanAPI/src/KlockanAPI.Infrastructure/Repositories/MeetingRepository.cs
@@ -71,7 +71,7 @@ public class MeetingRepository : IMeetingRepository
             {
                 UserId = userId,
                 ClassroomId = classroomId,
-                RoleId = user.RoleId,
+                RoleId = (int)user.RoleId,
                 Classroom = null,
                 User = null,
                 Role = null
